fix for pid auto

+++ Core/Inc/FlightController/PID.h
@@ -27,6 +27,6 @@ public:
 class PID_Controller
 {
 public:
-    static int16_t Update(PID *pid, float target, float measured, float dt);
-    static int16_t UpdateAngleWithGyroD(PID *pid, float targetAngleDeg, float measuredAngleDeg, float gyroDegPerSec, float dt);
+    static float Update(PID *pid, float target, float measured, float dt);
+    static float UpdateAngleWithGyroD(PID *pid, float targetAngleDeg, float measuredAngleDeg, float gyroDegPerSec, float dt);
 };
